Rename and simplify
