+++ examples/walking_issues.php
@@ -12,7 +12,7 @@ $api = getApiClient();
  *
  * @see https://developer.atlassian.com/static/rest/jira/5.0.html#id202584
  */
-$walker = new \chobie\Jira\Issues\Walker($api);
+$walker = new \Jira\Issues\Walker($api);
 $walker->push('project = TICKETACEG AND  updated > -1d ORDER BY priority DESC', '*navigable');
 
 // Okay, then just do foreach walker variable to pull issues.
