+++ src/Api/Client/ClientInterface.php
@@ -23,29 +23,38 @@
  * THE SOFTWARE.
  */
 
-namespace chobie\Jira\Api\Authentication;
+namespace Jira\Api\Client;
 
-interface AuthenticationInterface {
+use Jira\Api\Authentication\AuthenticationInterface;
 
-	/**
-	 * Returns credential string.
-	 *
-	 * @return string
-	 */
-	public function getCredential();
+interface ClientInterface {
 
 	/**
-	 * Returns user id.
+	 * Sends request to the API server.
 	 *
-	 * @return string
-	 */
-	public function getId();
-
-	/**
-	 * Returns password.
+	 * @param string $method Request method.
+	 * @param string $url URL.
+	 * @param array|string $data Request data.
+	 * @param string $endpoint Endpoint.
+	 * @param \Jira\Api\Authentication\AuthenticationInterface $credential Credential.
+	 * @param bool $is_file This is a file upload request.
+	 * @param bool $debug Debug this request.
 	 *
-	 * @return string
+	 * @throws \InvalidArgumentException When non-supported implementation of AuthenticationInterface is given.
+	 * @throws \InvalidArgumentException When data is not an array and http method is GET.
+	 * @throws \Jira\Api\Exception When request failed due communication error.
+	 * @throws \Jira\Api\UnauthorizedException When request failed, because user can't be authorized properly.
+	 * @throws \Jira\Api\Exception When there was empty response instead of needed data.
+	 * @return array|string
 	 */
-	public function getPassword();
+	public function sendRequest(
+		$method,
+		$url,
+		$data,
+		$endpoint,
+		AuthenticationInterface $credential,
+		$is_file = false,
+		$debug = false
+	);
 
 }
