+++ tests/Api/Client/CurlClientTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Tests\chobie\Jira\Api\Client;
+namespace Tests\Jira\Api\Client;
 
-use chobie\Jira\Api\Client\CurlClient;
+use Jira\Api\Client\CurlClient;
 
 class CurlClientTest extends AbstractClientTestCase {
 
 	/**
 	 * Creates client.
 	 *
-	 * @return \chobie\Jira\Api\Client\ClientInterface
+	 * @return \Jira\Api\Client\ClientInterface
 	 */
 	protected function createClient() {
 		return new CurlClient();
