+++ tests/Api/Client/PHPClientTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Tests\chobie\Jira\Api\Client;
+namespace Tests\Jira\Api\Client;
 
-use chobie\Jira\Api\Client\PHPClient;
+use Jira\Api\Client\PHPClient;
 
 class PHPClientTest extends AbstractClientTestCase {
 
 	/**
 	 * Creates client.
 	 *
-	 * @return \chobie\Jira\Api\Client\ClientInterface
+	 * @return \Jira\Api\Client\ClientInterface
 	 */
 	protected function createClient() {
 		return new PHPClient();
