+++ tests/resources/api_get_roles_of_project.json
